+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `set_preamble_length` method
 - `set_pa_boost`, `set_pa_rfo` methods
 - `frequency_error_indication_hz`, `frequency_error_indication_ppm` methods
+- `set_pa_ramp` method
 
 ## [0.1.0] - 2026-03-07